+++ src/script_trace/mod.rs
@@ -1,0 +1,6 @@
+mod callback;
+mod frame;
+
+pub use frame::{ScriptTraceFrame, ScriptTraceFrameKind, SigVersion};
+
+pub use callback::{set_script_trace_callback, unset_script_trace_callback, ScriptTracer};